Separating calendar into its own app
