+++ fec/cal/templates/calendar/calendar.html
@@ -1,0 +1,151 @@
+{% extends "base.html" %}
+{% load wagtailcore_tags %}
+{% load staticfiles %}
+{% load filters %}
+{% block body_class %}template-{{ self.get_verbose_name | slugify }}{% endblock %}
+
+{% block content %}
+
+{% include 'partials/breadcrumbs.html' with page=self style='secondary' %}
+
+<section class="main__content--full">
+  <div class="data-container__wrapper">
+    <div id="filters" class="filters">
+      <button type="button" class="filters__header js-filter-toggle">Edit filters</button>
+      <div class="filters__content" aria-hidden="true">
+        <div id="category-filters" class="js-accordion accordion--neutral" data-content-prefix="filter" data-open-first="true">
+          <div class="js-filter" data-filter="multi" data-name="category">
+            <button type="button" class="js-accordion-trigger accordion__button" id="label-1">Elections, deadlines and compliance</button>
+            <div class="accordion__content">
+              <div class="filter">
+                <fieldset class="js-filter js-dropdown" data-filter="checkbox">
+                  <legend class="label" for="state">State</legend>
+                  <ul class="dropdown__selected"></ul>
+                  <div class="dropdown">
+                    <button type="button" class="dropdown__button button--alt">Select</button>
+                    <div class="dropdown__panel" aria-hidden="true">
+                      <ul class="{{ css_class }}">
+                      {% for value, label in settings.CONSTANTS.states.items %}
+                      <li class="dropdown__item">
+                        <input id="state-{{ value }}" name="state" type="checkbox" value="{{ value }}">
+                        <label class="dropdown__value" for="state-{{ value }}">{{ label }}</label>
+                      </li>
+                      {% endfor %}
+                      </ul>
+                    </div>
+                  </div>
+                </fieldset>
+              </div>
+              <div class="filter">
+                <fieldset class="js-sub-filter" data-name="category" data-filter-label="label-1">
+                  <legend class="label">Election dates</legend>
+                  <ul>
+                    {% for value, label in settings.CONSTANTS.election_types.items %}
+                      <li>
+                        <input id="{{ value | clean_whitespace }}" type="checkbox" name="category" value="{{ value }}" />
+                        <label for="{{ value | clean_whitespace }}">{{ label }}</label>
+                      </li>
+                    {% endfor %}
+                  </ul>
+                  <span class="t-note">Includes primary, general and special elections as well as caucuses and conventions</span>
+                </fieldset>
+              </div>
+              <div class="filter">
+                <fieldset class="js-sub-filter" data-name="category" data-filter-label="label-1">
+                  <legend class="label">Filing deadlines</legend>
+                  <ul>
+                    {% for value, label in settings.CONSTANTS.deadline_types.items %}
+                      <li>
+                        <input id="{{ value | clean_whitespace }}" type="checkbox" name="category" value="{{ value }}" />
+                        <label for="{{ value | clean_whitespace }}">{{ label }}</label>
+                      </li>
+                    {% endfor %}
+                  </ul>
+                </fieldset>
+              </div>
+              <div class="filter filter--meeting">
+                <fieldset class="js-sub-filter" data-name="category" data-filter-label="label-1">
+                  <legend class="label">Reporting and compliance periods </legend>
+                  <ul>
+                    {% for value, label in settings.CONSTANTS.reporting_periods.items %}
+                      <li>
+                        <input id="{{ value | clean_whitespace }}" type="checkbox" name="category" value="{{ value }}" />
+                        <label for="{{ value | clean_whitespace }}">{{ label }}</label>
+                      </li>
+                    {% endfor %}
+                  </ul>
+                </fieldset>
+              </div>
+            </div>
+            <button type="button" class="js-accordion-trigger accordion__button" id="label-2">Meetings and outreach</button>
+            <div class="accordion__content">
+              <div class="filter">
+                <fieldset class="js-sub-filter" data-name="category" data-filter-label="label-2">
+                  <legend class="label">Commission meetings </legend>
+                  <ul>
+                    {% for value, label in settings.CONSTANTS.meeting_types.items %}
+                      <li>
+                        <input id="{{ value | clean_whitespace }}" type="checkbox" name="category" value="{{ value }}" />
+                        <label for="{{ value | clean_whitespace }}">{{ label }}</label>
+                      </li>
+                    {% endfor %}
+                  </ul>
+                </fieldset>
+              </div>
+              <div class="filter">
+                <fieldset class="js-sub-filter" data-name="category" data-filter-label="label-2">
+                  <legend class="label">Outreach </legend>
+                  <ul>
+                    {% for value, label in settings.CONSTANTS.outreach_types.items %}
+                      <li>
+                        <input id="{{ value | clean_whitespace }}" type="checkbox" name="category" value="{{ value }}" />
+                        <label for="{{ value | clean_whitespace }}">{{ label }}</label>
+                      </li>
+                    {% endfor %}
+                  </ul>
+                </fieldset>
+              </div>
+            </div>
+            <button type="button" class="js-accordion-trigger accordion__button" id="label-3">Legal</button>
+            <div class="accordion__content">
+              <div class="filter">
+                <div class="js-sub-filter" data-name="category" data-filter-label="label-3">
+                  <ul>
+                    {% for value, label in settings.CONSTANTS.rule_types.items %}
+                      <li>
+                        <input id="{{ value | clean_whitespace }}" type="checkbox" name="category" value="{{ value }}" />
+                        <label for="{{ value | clean_whitespace }}">{{ label }}</label>
+                      </li>
+                    {% endfor %}
+                  </ul>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+    <div class="data-container">
+      <div class="data-container__head calendar__head">
+        <h1 class="data-container__title js-calendar-title"></h1>
+        <div class="data-container__action js-calendar-action">
+          <div id="calendar-subscribe" class="export-widget"></div>
+          <div id="calendar-download" class="export-widget"></div>
+        </div>
+      </div>
+      <div class="data-container__widgets js-data-widgets">
+        <div class="js-filter-tags data-container__tags">
+          <div class="js-panel-controls panel__controls">
+            <button type="button" class="js-filter-clear filters__toggle filters__clear" aria-hidden="true">Clear all filters</button>
+          </div>
+        </div>
+      </div>
+      <div id="calendar"></div>
+    </div>
+  </div>
+</section>
+{% endblock %}
+
+{% block extra_js %}
+  <script type="text/javascript" src="{% static 'js/calendar-page.js' %}"></script>
+{% endblock %}
